Remove usage banner use
