+++ crates/agent/src/ui.rs
@@ -11,4 +11,3 @@ pub use agent_preview::*;
 pub use animated_label::*;
 pub use callout::*;
 pub use context_pill::*;
-pub use usage_banner::*;
